Finished exercise 4
Finished exercise 4 with the use of a outer for loop and an inner while loop.
The for loop traversed the multi-rest parameter arguement array that is created
when you may have multiple arguements. While the while loop traversed a copy of
the original array.

1. Traverse the multi-argument array. When traversing, the while loop will check if the current
element in multi-arugment array is in the array. We splice out the element in the array.