Completed Final JavaScript Exercise
Used if statements to solve the problem. Mainly i used the if statements
to help with determing the age of a person. The solution, provides a function
for getting back the age of the object. Next time will go with using a function
instead. The use of multiple if's made my code less readable.

-the reduce method for javascript was used.
ex: someArray.reduce(x,y)
    here x is a accumlator value, it will change based on the call back function,
    while y will iterate through someArray.